requirements.txt
openai==1.66.3
langgraph==1.0.9
PyYAML
zstandard
